[Add] #199 - 출석하기 관련 SampleData 추가
